fix dijikstra algo

+++ src/veins/Graph.cc
@@ -43,11 +43,11 @@ void Graph::addEdge(const Edge& edge) {
     edgeCount++;
 }
 
-size_t Graph::getNodeCount() const {
+int Graph::getNodeCount() const {
     return nodes.size();
 }
 
-size_t Graph::getEdgeCount() const {
+int Graph::getEdgeCount() const {
     size_t count = 0;
     for (const auto& kv : adjList) {
         count += kv.second.size();
